pyproject.toml
[tool.poetry]
name = "recommend-filter"
version = "0.1.0"
description = ""
authors = ["pjt3591oo <pjt3591oo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
scikit-surprise = "^1.1.1"
sklearn = "^0.0"
nltk = "^3.6.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
